requirements.txt
PySide6==6.8.3
SQLAlchemy==2.0.36
qtawesome
python-barcode
pillow
numpy>=1.26.0
reportlab
